Finalizing topUp system
