--- app/api/middleware/middleware.ts
@@ -1,8 +1,0 @@
-import nextConnect from 'next-connect'
-import multipartFormParser from './multipart-form-parser'
-
-const middleware = nextConnect()
-
-middleware.use(multipartFormParser)
-
-export default middleware
